More carefully look for compiled asdf.lisp.
Found that the current load recipe (load "../asdf") in the test
scripts was brittle, because the presence of asdf.fasl was obscured
if there was an "asdf/" directory in ..  Unfortunately, the result
of makeinfo --html asdf.texinfo created just such a directory!
fe[nl]ix kindly provided a pointer to compile-file-pathname that
solved the problem.