Fixed bugs in the tests.
There were bugs in the tests having to do with repeatedly calling
quit-on-error in a single test script.  Only one of those calls
will be evaluated, and then the script will exit.  That means that
multiple calls to quit-on-error masked failures in the tests.

Also fixed a package-related error in test1.script that I conjecture
led to the mistaken use of multiple calls to quit-on-error.