Putative test for module dependencies.
This test checks to see if a file in a module will be recompiled if a file
it depends on changes.  In "classic" asdf this does not happen --- it's
launchpad bug https://bugs.launchpad.net/asdf/+bug/502946

This is just a test for the condition.  The test fails in the current state
of ASDF.