Tentative fix for the module dependency bug.
I believe that this patch, using a dynamic variable, fixes the TRAVERSE
method so that the components of a module will be operated on if any of the
module's dependencies are activated in TRAVERSE.  However, this proposed
patch has ramifications.  Because SYSTEM isa MODULE, this means that
downstream systems will be compiled if upstream ones change, which ASDF has
never done before.