Updated the manual to address launchpad bug 610717.
Previously the manual directed users to override the source-file-type
method on cl-source-file.  I doubt that this was /ever/ a good idea, but it
certainly won't work any more, because the signature of the
source-file-type generic function has changed.  Killed the old discussion,
flagged it as having changed, and then added a FAQ entry.