2.016.1: Fixes for ABCL, GCL; tweaks for LispWorks and CCL.
Use probe-file for probe-file* in LispWorks.

Make ABCL's jar translation happier with
the recent compile-file-pathname* change for Stelian by redoing cfp myself.

Make truenamize work when the merged pathname is relative.
Makes it happier when e.g. the *default-pathname-defaults* is #p"",
which is the case notably on CCL.

Try and fail to support GCL 2.6 and have 2.7 fail less;
too many bugs, in CONDITIONs, PATHNAMEs, and more,
plus compiler bug that breaks coerce-pathname.