2.26.59: Simplify TRAVERSE signature. Fix subtle bug in VISIT-DEPENDENCIES.
The TRAVERSE issue makes it easier for user code to override TRAVERSE:
don't include FORCE FORCE-NOT VERBOSE in mandatory keyword arguments
for all methods. A method is notably provided by buildapp.

The bug in VISIT-DEPENDENCIES was causing IF-FEATURE'ed out
components to flag whoever depended on them as requiring to be rebuilt.
This notably caused a legacy SB-ROTATE-BYTE to always want to rebuild.
The dependency visit is now predicated on VALID-ACTION-P,
which requires VISIT-DEPENDENCIES to take an additional PLAN argument.