2.26.93: play nicer with packages, notably for Allegro and CLISP
On Allegro, frob the autoload earlier and intern the gethostname later
so we don't unnecessarily autoload stuff.

On CLISP, don't undefine functions, just unintern everything in the beginning,
and it won't complain about disappearing methods anymore.

Also, only retrigger the upgrade attempt on source-registry change
if ASDF was not upgraded yet. This avoids unnecessary reloading of ASDF.

Finally, be sure to share more symbols between everyone - hopefully,
all the symbols that were used in ASDF and got recycled.
Slot names are a big one, notably.