2.26.106: a new primary/secondary convention for subsystems and more
A system named foo/bar will be looked up in a registered file foo.asd.
This makes it possible to sensibly name several systems in a .asd file
and still have asdf find them all by name.
We now use it to not have the asdf header be part of asdf-driver.

A new generic operation build-op makes it possible to deal with
the fact that the default operation for systems need not be load-op,
yet that the end-user does not want to have to know the exact operation
for each of the system he uses.

ensure-pathname has been improved again, and its clients tweaked.

remove-keys was made more sensible by using EQL rather than STRING-EQUAL.

:version need not take the first form in a file, but can take
a path to a subform. This makes it possible to get the exact subform
in a Lisp source file that has the string.
See SUB-OBJECT for how these paths work.

Cleanup of the semantics of ENSURE-FUNCTION on CONSes: use APPLY, not EVAL.