2.26.146: restore W-C-U around perform-plan for backward compatibility.
Also, move entry-point into system and remove class bundle-system.
I was tempted to have test-op call it, but then it's a problem
when the base system has a non-test entry-point, so we probably need
a separate test-entry-point slot.

Update the default *asdf-debug-utility* to do more of the right thing.

Update changelog.

Tweak package to avoid triggering a bug in ABCL trunk.