2.29.2: CCL: punt on upgrade from ASDF2. load-asd: use global readtable.
Package upgrade on CCL still fails due to setf function issues.
Until they are resolved, punt if upgrading from earlier than 2.27.

iolib 0.7.3 modifies the global readtable, and is thus incompatible with
the with-standard-io-syntax in load-asd, unless we re-bind readtable
to preserve this dirty global thing. Until we fix iolib and all clients,
revert to higher compatibility.