2.29.9: in load-asd, preserve *print-pprint-dispatch* as well as *readtable*.
Although I'd like to isolate .asd files from each other,
copy-readtable and copy-pprint-dispatch are slightly too expensive
to use lightly (cons 1648 and 2192 bytes respectively on CCL 1.9 x86-64,
and more like 5300 and 10300 bytes on SBCL 1.1.3 x86-64).
There is no portable way to ensure the tables are read-only, and
bad things can happen if they are mutated while bound to the standard defaults.
So for now, I'm doing the backward-compatible thing of letting users
take responsibility for binding and mutating them, or more likely for success,
avoiding to bind and mutate them.

Thanks once again to Anton Vodonosov for his help with cl-test-grid.