2.31.5: Fix deferred-warnings on CCL. Fix old FEATURE broken by 2.31.2.
On CCL, we were trying to persist data that can't be written and read back
regarding the specializers of context function and values of arguments.
Now drop the information at the cost of slightly less precise warning messages.

In 2.31.2, we failed to export FEATURE from asdf/operation,
causing this compatibility feature to fail.