Fixes for UIOP:RUN-PROGRAM on ACL.
On ACL, the :INPUT and :OUTPUT arguments to EXCL:RUN-SHELL-COMMAND take
NIL for the behavior of inheriting input and output handles from the
Lisp process.  These were encoded as T in UIOP:RUN-PROGRAM, instead of
NIL, causing ACL to get errors trying to open T as a file designator.

Added a couple of test cases to test-run-program.script to illustrate
the problem.

Note that this fix does not provide the prescribed behavior for :OUTPUT
NIL (which should take the program's output and throw it on the floor).
There is a TODO in uiop:run-program explaining what's missing.  In
brief, ACL's EXCL:RUN-SHELL-COMMAND does not support this behavior out
of the box, so one would have to capture the program's output, and then
throw it on the floor.