Gracefully handle the case where a package-system is missing a package.
This bug was found while testing with ECL, which somehow
evaluated the .asd twice during test-bundle.script
when other implementations evaluated it only once;
which revealed that the second time around,
the package-system would override the explicit secondary definitions
of test-asdf/dll-test and test-asdf/dll-user, and then
would fail ungracefully due to the absence of defpackage in these files.
That latent bug was present in all implementations,
but not stressed by the test system.