Remove compatibility stubs for old bundle op classes and document.
The compatibility stubs were not fully functional, so rather than
expend a lot of time fixing the stubs, we have done away with the old
names for the bundle ops.

Added documentation for this in the manual, as a new FAQ.

While indexing the class names, I found that we incorrectly used findex
for classes, instead of tindex.  Added syncodeindex so that we can use
the tindex command going forward without breaking everything.