Fix chdir, getcwd, run-program on ABCL, add tests.
The JVM doesn't give access to native chdir without some extensions that
aren't available with ABCL. There's a user.dir system property whereby
some Java programs kind of fake it, but ABCL doesn't play nice with it.
Therefore, use *default-pathname-defaults* as our "current directory",
and be sure to chdir to it before to run commands with run-program.
Try to make the run-program chdir feature work on Windows.