uiop: gut the last wrapping in make-pathname*
I initially introduced all that make-pathname* madness on the road to
ASDF 2.27, back when I was trying to normalize pathname components so
I may have pathname-indexed entries in the asdf-cache (to minimize
costly filesystem access from file-write-date and such, that could
also introduce inconsistency in the build plan).

Problem is, there was no good way (even non-portable) to normalize
pathnames in a way that works when you use all of parse-namestring,
make-pathname, directory, probe-file, translate-logical-pathname,
merge-pathnames, etc. In the end (as late as b4dc78d, before 3.0.2.13!),
I conceded defeat, for it was an uphill battle, and instead asdf-cache
is indexing things using namestrings instead, which is more stable (see
asdf/cache:normalize-namestring).

The hacks in make-pathname* are therefore not needed anymore,
and the function can be done away with.

This commit guts the contents of make-pathname* and marks it deprecated.
Just use make-pathname instead. It also updates uiop and asdf to not
use make-pathname* but make-pathname. Some future commits may move it to
backward-driver and eventually start issuing warnings of some sort when
it's used.