Add tests for defsystem-depends-on
Attila offered these tests, that currently fail at HEAD,
for which I have a patch (to be included in next commit).
They check that an error when loading defsystem-depends-on dependencies
doesn't cause an incomplete system to be registered but never properly
parsed, that thereafter pollute the build.

A real solution would imply to fix lp#1500578 though:
   ASDF should handle change in defsystem-depends-on dependencies
   https://bugs.launchpad.net/asdf/+bug/1500578