Define backward load-sysdef to unbreak swank-asdf
Undelete function load-sysdef, which must continue to exist as long as
people use the current version of SLIME (to be fixed imminently as of
2015-12-01), though it is OK for it to just issue an error.

Indeed, swank-asdf supposes that any backward-compatibility function it
defines will exist forever, by unconditionally calling fdefinition on
them. Now, it depends on load-asd, the backward-compatible definition of
which (for asdf older than 2.26.149) depends on load-sysdef, that it also
defines for versions older than 2.16.