Fix directory-files to be linear not quadratic
Bug found by @axionic on github: on very large directories,
directory-files was slow, because it was using remove-duplicates
with non-hashable test function #'pathname-equal.
But this functionality was only really needed for the purpose
of dealing with logical-pathnames, and having a few duplicates
in case of wanton symlinks isn't much of an issue.