Modify test-running scripts to catch SERIOUS-CONDITIONs.
There are some serious conditions (e.g., STORAGE-CONDITION) that are not
ERRORs.  Previously, our test scripts only caught ERRORs, meaning that
if a lisp implementation encountered a non-ERROR SERIOUS-CONDITION in
the course of running the tests, odd behavior could result.  E.g.,
running in ECL we would drop into the debugger instead of exiting with a
non-zero exit status.