Redefined *FATAL-CONDITIONS* per 1605650.
See https://bugs.launchpad.net/asdf/+bug/1605650

According to the spec, SERIOUS-CONDITIONs are "All conditions serious
enough to require interactive intervention if not handled should inherit
from the type serious-condition."  This is almost exactly the definition
of *FATAL-CONDITIONS*, so I substituted SERIOUS-CONDITION for ERROR
here.

Added exceptions to the set of *FATAL-CONDITIONS* in order to fix test
condition handling for CCL.  The exceptions permit us to have a class of
conditions (e.g., SERIOUS-CONDITION) that has a subclass that is not a
FATAL-CONDITION, but that does not have an obvious more specific
alternative.

The specific example that caused me to add this: On shutdown, CCL
signals CCL:PROCESS-RESET, a subclass of SERIOUS-CONDITION. So recent
test harness modifications to handle SERIOUS-CONDITIONs caused the test
scripts to mistakenly think there were failures because of calls to
EXIT-LISP.