Bug fix: Exit code with LispWorks 7
By having :process always be a stream on LispWorks 7, we make sure
that the process can be killed, have its vital signs checked, and its
PID obtained. In a simple and consistent manner.

With LispWorks 7, pid-exit-status is no longer available. We need a
stream to obtain the exit status. With save-exit-status, we can be
sure that we're given one.