Change :if-*-exists default to :supersede
The difference between :overwrite (the former default) and :supersede
is only that the latter should not lead to an error if the file does
not exist. There is plenty of support for :supersede behaviour (not
necessarily under that name) but not for :overwrite behaviour. Hence
we make it the default and add a normaliser.

Since normalisation is currently only necessary for CLISP and CLISP
only has :if-output-exists support (not :if-error-output-exists or
:if-input-does-not-exist), we only translate :if-output-exists for now.