Fix functioning under Cygwin.
Under Cygwin, the CL_SOURCE_REGISTRY will be a set of *cygwin*
pathnames. But when this gets passed to the Windows lisps (and through
them to other programs), that caused failures.
So now we check for cygwin, and if we find it, we use the Cygwin
"cygpath" command to translate the CL_SOURCE_REGISTRY into Windows
native paths.
Possible problem: originally the CL_SOURCE_REGISTRY in the Makefile was
terminated with ":" which at least sometimes was treated as an implicit
".".  But then I found that this caused failures on some Windows tests,
because that was a relative pathname, which the path translator didn't
like.