Update the web page
Note that the latest release is 3.1.7. Also note that since CLISP adopted it
in May 2016, all active Common Lisp implementations provide ASDF 3.1. Yay.

Also import some information that used to be in the README.md but removed
in 81452d5. That README also had information on building and testing with
the asdf-tools, and on the directory structure of the git repo.