Restore information in the README.md
Robert deleted most information in the README.md in 81452d5.
But while some of it was indeed redundant with the web page and manual
and a pain to maintain, most of it was not covered at all in them,
and pertained to how to use the source code repository, for new developers.
I restored that information.