bundle: component-depends-on: simplify hairy code
Create a local function linkable-system, which given a system in the current
image, returns an equivalent linkable system.

This part of code was also buggy, because if it did found `uiop', it
didn't include `asdf', but if none was found, it included `uiop' *and*
`asdf' unconditionally. This inconsistent behaviour was fixed and now we
try to inject `uiop', and if not found – `asdf'.