Simplify TOUCH-FILE.
There was a very complex conditional in the middle controlling how it
was implemented.  I simplified the conditions to make it easier to
understand how it works.
I also modified it so that if you tell TOUCH-FILE to use the cache, and
the cache hasn't been initialized, it will raise an error.