Move auto-version from immutable to preloaded systems
Daniel Kochmanski convinced me that immutable systems are "just" blessed
preloaded systems; the automatic version feature of immutable systems can thus
be generalized to all preloaded systems.

Update several docstrings.

Also, document that direct access to *immutable-systems* is not supported.
Contact the maintainers if you need an interface to that.