Avoid duplicate call to lisp-implementation-version
The way that lisp-version-string is written, it will fall back to
lisp-implementation-version unless a platform-specific override is
provided. Since the body of lisp-version-string uses

  (car (list #+foo exp1 exp2))

rather than

  (or #+foo exp1 exp2)

it is theoretically possible that lisp-version-string would return NIL
if that's what exp1 evaluates to, which would not also be returned by
lisp-implementation-version, which it previously would not have.

Such behaviour would qualify as a bug in lisp-version-string, though.