Fix the upgrade hooks
The order in which the upgrade fixups were called wasn't very robust.
We need to first get a list of systems being defined, then
clear defined systems (but not asdf, and *neither uiop* that asdf depends on),
then upgrade the configuration, and *finally* reload any of the systems being
defined (including their defsystem-depends-on dependencies!),
which means this all has to be coordinated within the same function.