Simplify operation-forced
Don't use operation-original-initargs in operation-forced.
operation-original-initargs is deprecated, and so is operation-forced.
The only, misguided, user of operation-forced is swank.asd,
which conditionally loads the code only if the operation is forced;
but the perform method that does the check will only be called twice
if the operation is forced, anyway, so the correct answer is T.
(And the first time around, it's OK to return T, too.)