Stop declaring arguments ignorable
DIRECTORY is important. It was never actually ignored, except on Allegro,
on which the implementation itself handles the option. Make that explicit.

The other arguments are always passed... though it's unclear whether they
are heeded in all implementations, since we have :allow-other-keys t.
That deserves investigation and a comment update, at some point.

Also, tweak some comments.