Get fully rid of operation initargs
ASDF never supported operation initargs, and its component-operation-times
always assumed that two operations of the same class were equivalent.
Make it explicitly so and enforce it.

Do NOT pass keys from operate to make-instance operation.
Catch any attempt to create an operation with initargs and raise an error.
Make sure no operation class has any slot that isn't :allocation :class
(with some MOP magic we could do it in at class definition time,
but that's not portable enough and we don't want to reimplement closer-mop
or depend on it.)

This is a preliminary to allowing CONS instead of NODE-FOR to identify actions,
whether or not we want to allow some form of initargs in the future.

Remove unused COMPILE-OP-FLAGS and COMPILE-OP-PROCLAMATIONS.
Remove MAKE-BUILD that depended on the unsupported operation initargs.
Use PROGRAM-OP on a PROGRAM-SYSTEM instead.
Remove the confusingly misnamed and not-so-useful BUILD-SYSTEM function;
it can be done better in userland.