Fix bundle-op and concatenate-source-op file order
Fix direct-dependencies, that was reversed from file order,
leading to inversions in the order of input files for bundle-op and
concatenate-source-op as compared to what a user would expect.

Check all the bootstrap file lists and provide an *actionable* error message
when discrepancies are detected.

Also fix bundle's use of required-components to be in line with an old comment.
And remove the comment.