Restore proper use of required-components, with explanation
Undo part of 0e0a851b, that changed the keep-operation in invocations of
required-components from compile-op to load-op. Explain why in a comment of
bundle.lisp: so any compiled file for the component itself be included,
which notably matters when using a package-inferred-system.

This notably fixes test-bundle.script, test-program.script on mkcl and ecl,
and test-concatenate-source.script on ecl.

Also simplify the test-bundle support files and remove some warnings.