Fix monolithic-dll-op on ECL
Use required-components in a way that is more robust when
*load-system-operation* is load-bundle-op rather than load-op,
by using basic-compile-op and/or basic-load-op as the :keep-operation
rather than compile-op or load-op respectively.

Test-bundle was passing the first time around, but not the second time.
The first time, it was only passing because the appropriate file were
loaded while building the dependencies of the target monolithic dll;
but monolithic-dll-op was already buggy and not linking all these dependencies.
The second time around, there was no building step and the proper dependencies
had failed to be linked in, so the loading failed.

This fixes https://bugs.launchpad.net/asdf/+bug/1626994