Fix test-run-program on windows
We cannot use `exit` on Windows with :force-shell nil (it's a builtin);
use `cmd /c exit` instead.

Using `exit` on Unix with :force-shell nil may or may not work (it's a
builtin but CL implementations typically invoke a shell anyway).
`true` and `false` are guaranteed to exist by the POSIX standard.