Stop using .a's as input for anything but image
There is no robust portable way to combine multiple .a files together;
libtool kind of works but has weird limitations and is a cumbersome dependency
to require, and even it doesn't work well in cross-compilation setting.
The ECL maintainer Daniel K suggests we should always link from object files:
https://bugs.launchpad.net/bugs/1647391

This patch still uses static libraries as input for images,
but even that could be changed.