Stop warning in define-convenience-methods explain
The convenience methods recursively call explain, which cause a style-warning
if the compiler-macro is already defined, and a different one if it is defined
afterwards. The solution is to declare the function notinline in the body of
the convenience methods.

This fixes test-clean-load on sbcl, and more generally
hushes loading asdf.lisp from source.