Fix images for ASDF-dependent systems on ECL
Recent ECLs have libasdf.a instead of asdf.a.

Not all bundle-op is a basic-compile-op (e.g. concatenate-source-op).
Declaring otherwise was causing asdf/defsystem to appear as a compiled
dependency, causing incorrect double inclusion of symbols in the link
when asdf was explicitly depended on.

Make sure to include only one of uiop or asdf as the fallback support file,
and only when neither uiop nor asdf is explicitly depended on.

Improve test-program.script for ECL.

This fixes https://bugs.launchpad.net/bugs/1660547