plan: implement phase level escalation
Track at which level of nested calls to OPERATE a not-done action is needed;
if it is needed at a higher level than previously traversed, then traverse again
to schedule it (and its transitive dependencies) to be done in the higher level.
This fixes cases, such as in (test-system "cl-protobufs-tests"), where a library
was already marked as needed at a low level of OPERATE, but if actually needed
at higher level due to a defsystem-depends-on, and was previously not loaded
early enough (after this plan refactoring).