Fix bug detecting need to reload defsystem.
Previously, if one bumped a version number that was included into a
system definition using :READ-FILE-FORM or :READ-FILE-LINE, ASDF would
not notice the change (so it could, for example, fail to recognize that
a depended-on system has had its version number bumped and now satisfies
a new :VERSION requirement).

This was fixed by adding ADDITIONAL-INPUT-FILES to the object model.
These serve the purpose of giving a place other than INPUT-FILES for ASDF
to stash input files that it "notices" on its own (e.g., when processing
:READ-FILE-<FOO>).  ASDF can't stash such files in INPUT-FILES, because
it must leave the INPUT-FILES primary methods open for users to
override.