ecl: create-image: don't call internal function compute-init-name
It is wrong to call this function. If :init-name is not supplied, or
NIL ECL will compute the init name itself (it is not a new
mechanism).

Enabling specifying init-name makes it possible to build static and
shared libraries with known initialization function (important, when
CL module is linked with bigger C/C++ project outside ASDF).

If :init-name is NIL, it will default to compute-init-name
internally. Calling explicitly is not guaranteed to work in the
future (internal mechanism).