Fix upgrade issues on ABCL and ECL
Fix upgrade on ABCL, that failed to properly process the fact that
initialize-source-registry had moved from asdf/find-system to
asdf/source-registry, yet was called in a cleanup function.

Fix upgrade on ECL, that failed to correctly process that
find-component was fmakunbound.
It also seems to clear *central-registry* in some upgrade scenarios
(that also moved from find-system to system-registry).

Tweak script-support to not unconditionally frob-packages
when ASDF isn't loaded yet, and to not include the asdf directory
when loading and testing only UIOP.