Preserve syntax tables in perform define-op
ASDF 3.3.0 introduced an unwanted discrepancy from 3.2.1, whereby
the readtable and print-pprint-dispatch table were not preserved
while loading a .asd file, which caused bad libraries like puri
to try to corrupt the standard readtable when loaded as
defsystem-depends-on dependency or equivalent.
See bug report on the asdf-devel mailing-list.
The real solution is to redo the syntax-control branch.