utility: remove fwrap, table-alist
These utilities are not currently used; they can be reinstated later if needed.

fwrap was for composing a list of wrapper functions, except that the last was
the special case function to be wrapped. Better to just (define and) use
compose, with funcall (or call-function) as the base case.

table-alist was as the name implies, and was intended to help go between
internal table representations and external alist representations of
per-system variables.