REMOVE support for per-system special bindings
Undo all changes that implement per-system special bindings, whereby a set of
syntax variables are set to known values around actions concerning a system.
This reversal is made in view of making a minimal change towards syntax control,
and may be reversed again in some future.

A rollback of this commit is expected at some indefinite point in the future.