REMOVE support for basic-prepare-op, cl-reading-op
Undo all changes that implement basic-prepare-op, cl-reading-op, and
binding of syntax around perform'ance of cl-reading-op.
Note that these changes depend on with-action-context, which was reverted
in the previous commit regarding per-system variable bindings.
These reversals are made in view of making a minimal change towards
syntax control, and may be reversed again in some future.

A rollback of this commit is expected at some indefinite point in the future.